resume_analyzer/requirements.txt
langchain
langchain-openai
langchain-google-genai
google-generativeai
gradio
pdfminer.six
nltk
python-dotenv
